requirements.txt
alabaster==0.7.10
attrs==17.3.0
awscli==1.14.6
Babel==2.5.1
bleach==2.1.2
bokeh==0.12.11
botocore==1.8.10
Bottleneck==1.2.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
cmocean==1.1
colorama==0.3.7
configobj==5.0.6
coverage==4.4.2
cycler==0.10.0
dask==0.16.0
decorator==4.1.2
distributed==1.20.2
docutils==0.14
entrypoints==0.2.3
flake8==3.5.0
gitdb2==2.0.3
GitPython==2.1.7
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
hypothesis==3.38.5
idna==2.6
imagesize==0.7.1
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyter-dashboards==0.7.0
jupyter-dashboards-bundlers==0.9.1
line-profiler==2.1.2
locket==0.2.0
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
memory-profiler==0.47
mistune==0.8.3
msgpack-python==0.4.8
nbconvert==5.3.1
nbdime==0.4.0
nbformat==4.4.0
netCDF4==1.3.1
notebook==5.3.1
numpy==1.13.3
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
pexpect==4.3.1
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psutil==5.4.1
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pykdtree==1.2.2
Pympler==0.5
pyparsing==2.2.0
pyproj==1.9.5.1
pyresample==1.8.1
pytest==3.3.1
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
rsa==3.4.2
runipy==0.1.5
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.3
snowballstemmer==1.2.1
sortedcontainers==1.5.7
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
spiceypy==2.1.0
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
ttide===0.3-exp
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.3
xarray==0.10.0
zict==0.1.3
zope.interface==4.4.3
